+++ CSharp/DevVmPowershell/Helpers.Tests.Integration/Tests/AgentHelperTests.cs
@@ -20,8 +20,7 @@ namespace Helpers.Tests.Integration.Tests
 				sqlAdminUserName: TestConstants.SQL_USER_NAME,
 				sqlAdminPassword: TestConstants.SQL_PASSWORD);
 			IRestHelper restHelper = new RestHelper();
-			Sut = new AgentHelper(connectionHelper, restHelper, TestConstants.RELATIVITY_INSTANCE_NAME,
-				TestConstants.RELATIVITY_ADMIN_USER_NAME, TestConstants.RELATIVITY_ADMIN_PASSWORD);
+			Sut = new AgentHelper(connectionHelper, restHelper);
 		}
 
 		[TearDown]